ripped out box2d and PID system

+++ geometry.h
@@ -7,9 +7,9 @@
 
 struct pos 
 {
-	float x = 0;
-	float y = 0;
-	float rot = 0;
+	double x = 0;
+	double y = 0;
+	double rot = 0;
 
 	int distance (const pos& p) const
 	{
